+++ bits/61_fcommon.js
@@ -16,8 +16,19 @@ var rc_to_a1 = (function(){
 	};
 })();
 
-/* no defined name can collide with a valid cell address A1:XFD1048576 ... except LOG10! */
 var crefregex = /(^|[^._A-Z0-9])([$]?)([A-Z]{1,2}|[A-W][A-Z]{2}|X[A-E][A-Z]|XF[A-D])([$]?)([1-9]\d{0,5}|10[0-3]\d{4}|104[0-7]\d{3}|1048[0-4]\d{2}|10485[0-6]\d|104857[0-6])(?![_.\(A-Za-z0-9])/g;
+var a1_to_rc =(function(){
+	return function a1_to_rc(fstr, base) {
+		return fstr.replace(crefregex, function($0, $1, $2, $3, $4, $5, off, str) {
+			/* TODO: handle fixcol / fixrow */
+			var c = decode_col($3) - base.c;
+			var r = decode_row($5) - base.r;
+			return $1 + "R" + (r == 0 ? "" : "[" + r + "]") + "C" + (c == 0 ? "" : "[" + c + "]");
+		});
+	};
+})();
+
+/* no defined name can collide with a valid cell address A1:XFD1048576 ... except LOG10! */
 function shift_formula_str(f/*:string*/, delta/*:Cell*/)/*:string*/ {
 	return f.replace(crefregex, function($0, $1, $2, $3, $4, $5, off, str) {
 		return $1+($2=="$" ? $2+$3 : encode_col(decode_col($3)+delta.c))+($4=="$" ? $4+$5 : encode_row(decode_row($5) + delta.r));